// src/shared/enums/QuestionCategory.ts
/**
 * @file QuestionCategory.ts
 * @author Mark Hutchison
 * An enum that details the exact options for the challenge die to roll.
 */

/**
 * An enum detailing the exact question categories for the challenge die.
 */
export enum QuestionCategory {
  TakeThreeAllPlay = 0,
  MusicalAllPlay = 1,
  MiscellaneousAllPlay = 2,
  Consequence = 3,
  TakeThreeMyPlay = 4,
  MusicalMyPlay = 5,
  MiscellaneousMyPlay = 6,
  ConsequenceB = 7,
}